factoring locreator and resolve functions to separate modules, using new locatex array arguments interface
