+++ lib/locreator.js
@@ -1,0 +1,79 @@
+'use strict';
+
+var async = require('async');
+
+module.exports = function locreator(nemo, locatorId, locatorJSON, parentWebElement) {
+  var locatorObject = {};
+
+  locatorObject[locatorId] = function () {
+    return nemo.drivex.find(locatorJSON, parentWebElement);
+  };
+  locatorObject[locatorId + 'By'] = function () {
+    return nemo.drivex.by(locatorJSON);
+  };
+  locatorObject[locatorId + 'Present'] = function () {
+    return nemo.drivex.present(locatorJSON, parentWebElement);
+  };
+  locatorObject[locatorId + 'Wait'] = function (timeout, msg) {
+    function waitReturnElement() {
+      return nemo.drivex.waitForElement(locatorJSON, timeout || 5000, msg).then(function() {
+        return nemo.drivex.find(locatorJSON, parentWebElement);
+      });
+    }
+    var wep = new nemo.wd.WebElementPromise(nemo.driver, waitReturnElement());
+    return wep;
+  };
+  locatorObject[locatorId + 'WaitVisible'] = function (timeout, msg) {
+    function waitVisibleReturnElement() {
+      return nemo.driver.wait(function () {
+        return nemo.drivex.visible(locatorJSON);
+      }, timeout || 5000, msg).then(function() {
+        return nemo.drivex.find(locatorJSON, parentWebElement);
+      });
+    }
+    var wep = new nemo.wd.WebElementPromise(nemo.driver, waitVisibleReturnElement());
+    return wep;
+  };
+  locatorObject[locatorId + 'Visible'] = function () {
+    return nemo.drivex.visible(locatorJSON, parentWebElement);
+  };
+  locatorObject[locatorId + 'OptionText'] = function (optionText) {
+    var d = nemo.wd.promise.defer();
+    nemo.drivex.find(locatorJSON, parentWebElement).then(function (selectEl) {
+      selectEl.findElements(nemo.wd.By.css('option')).then(function (elts) {
+        var current = 0;
+        var total = elts.length;
+        var found = false;
+        async.whilst(
+          function () {
+            return current < total;
+          },
+          function (callback) {
+            var elt = elts[current++];
+            elt.getText().then(function (txt) {
+              if (txt === optionText) {
+                found = elt;
+              }
+              callback();
+            });
+          },
+          function (err) {
+            if (found !== false) {
+              d.fulfill(found.click());
+            } else {
+              d.reject(new Error('couldn\'t find option with text: ' + optionText));
+            }
+          }
+        );
+      });
+    });
+    return d;
+  };
+  locatorObject[locatorId + 'OptionValue'] = function (optionValue) {
+    //select an option by its text (works only for Select elements)
+    nemo.drivex.find(locatorJSON, parentWebElement).then(function (selectEl) {
+      return selectEl.findElement(nemo.wd.By.css('option[value=' + optionValue + ']')).click();
+    });
+  };
+  return locatorObject;
+};
